+++ cli/scripts/util.py
@@ -16,6 +16,7 @@ import api, meta
 
 ONE_DAY = 86400
 ONE_WEEK = ONE_DAY * 7
+DEFAULT_PG = "16"
 
 bad_os_warn = False