+++ RELEASE-NOTES.md
@@ -18,6 +18,7 @@
   - add support for plv8-3.2.0 for dev and test
   - design for spock.db_create()
   - WIP Windoze CLI
+  - WIP refactor install-pgedge.py to use spock.db_create()
 
 
 ## done for 23.127 as of 2023-08-10